+++ tests/node-traversal.rs
@@ -57,7 +57,7 @@ fn test_first_element_child_edge_cases() {
 
 #[cfg_attr(not(target_arch = "wasm32"), test)]
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
-fn test_all_descendants() {
+fn test_descendants_iter() {
     let doc: Document = ANCESTORS_CONTENTS.into();
 
     let ancestor = doc.select("#great-ancestor");
@@ -78,6 +78,43 @@ fn test_all_descendants() {
         "first-child",
         "second-child",
         "grand-parent-sibling",
+    ];
+    assert_eq!(descendants_id_names, expected_id_names);
+}
+
+
+#[cfg_attr(not(target_arch = "wasm32"), test)]
+#[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+fn test_all_descendants() {
+    let doc: Document = ANCESTORS_CONTENTS.into();
+
+    let ancestor = doc.select("#great-ancestor");
+    assert!(ancestor.exists());
+
+    let ancestor_node = ancestor.nodes().first().unwrap();
+
+    // if you want to reuse descendants then use `descendants` which returns a vector of nodes
+    let descendants = ancestor_node.descendants();
+
+    // Descendants include not only element nodes, but also text nodes.
+    // Whitespace characters between element nodes are also considered as text nodes.
+    // Therefore, the number of descendants is usually not equal to the number of element descendants.
+    assert_ne!(descendants.len(), 5);
+
+    // with no depth limit
+    let descendants_id_names = descendants
+        .iter()
+        .rev()
+        .filter(|n| n.is_element())
+        .map(|n| n.attr_or("id", "").to_string())
+        .collect::<Vec<_>>();
+
+    let expected_id_names = vec![
+        "grand-parent-sibling",
+        "second-child",
+        "first-child",
+        "parent",
+        "grand-parent",
     ];
     assert_eq!(descendants_id_names, expected_id_names);
 }